third_party/blink/web_tests/http/tests/websocket/invalid-continuation-expected.txt

Test whether WebSocket aborts the connection when it receives an unexpected continuation frame without any leading text or binary frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
